feat(slack): split user and bot providers
