+++ src/oauth/oauth-credential-refresh-service.ts
@@ -2,7 +2,7 @@ import type { ResolvedCredential } from "../core/types.ts";
 import type { OAuthClientConfigService } from "./oauth-client-config-service.ts";
 
 import { ConnectionError } from "../connection-service.ts";
-import { refreshSlackOAuthCredential } from "../providers/slack/oauth.ts";
+import { getSlackOAuthTokenKind, normalizeSlackOAuthCredential } from "../providers/slack/oauth.ts";
 import { readOAuthClientConfigMetadata } from "./oauth-client-config-service.ts";
 import { expiresAtFromLifetime, requestRefreshToken } from "./oauth-token.ts";
 
@@ -45,15 +45,11 @@ export class OAuthCredentialRefreshService implements IOAuthCredentialRefresher 
         tokenUrl: this.clientConfigs.resolveEndpointUrl(service, auth.refreshTokenUrl ?? auth.tokenUrl, config),
         createError: (message) => new ConnectionError("oauth_token_refresh_failed", message),
       });
-    if (service == "slack") {
-      return refreshSlackOAuthCredential(credential, requestTokenRefresh);
-    }
-
     const refreshed = await requestTokenRefresh(credential.refreshToken ?? "");
     const expiresIn =
       refreshed.expiresAt === undefined ? credential.metadata.expires_in : refreshed.metadata.expires_in;
 
-    return {
+    const refreshedCredential: OAuthCredential = {
       ...refreshed,
       refreshToken: refreshed.refreshToken ?? credential.refreshToken,
       // `expires_in` is optional on a refresh response, and a credential without an
@@ -72,5 +68,9 @@ export class OAuthCredentialRefreshService implements IOAuthCredentialRefresher 
         refreshedAt: new Date().toISOString(),
       },
     };
+    const slackTokenKind = getSlackOAuthTokenKind(service);
+    return slackTokenKind
+      ? normalizeSlackOAuthCredential(refreshedCredential, slackTokenKind, "oauth_token_refresh_failed")
+      : refreshedCredential;
   }
 }
