+++ src/oauth/oauth-flow-service.test.ts
@@ -9,6 +9,7 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { createCatalogStore } from "../catalog-store.ts";
 import { ConnectionService } from "../connection-service.ts";
 import { provider as slackProvider } from "../providers/slack/definition.ts";
+import { provider as slackbotProvider } from "../providers/slackbot/definition.ts";
 import { AesGcmSecretCodec } from "../server/secrets/secret-codec.ts";
 import { OAuthClientConfigService } from "./oauth-client-config-service.ts";
 import { OAuthFlowService } from "./oauth-flow-service.ts";
@@ -381,51 +382,79 @@ describe("OAuthFlowService", () => {
     await expect(services.connections.getCredential("example")).resolves.toBeUndefined();
   });
 
-  it("stores Slack's user grant outside token metadata", async () => {
-    const services = createServices([{ ...slackProvider, actions: [] }]);
-    await services.clientConfigs.upsertConfig({
-      service: "slack",
-      clientId: "client-id",
-      clientSecret: "client-secret",
-    });
-    vi.stubGlobal(
-      "fetch",
-      vi.fn(async () =>
-        Response.json({
+  it("uses separate Slack user and bot authorization paths with the same OAuth app", async () => {
+    const services = createServices([
+      { ...slackProvider, actions: [] },
+      { ...slackbotProvider, actions: [] },
+    ]);
+    for (const service of ["slack", "slackbot"]) {
+      await services.clientConfigs.upsertConfig({
+        service,
+        clientId: "shared-client-id",
+        clientSecret: "shared-client-secret",
+      });
+    }
+    const fetcher = vi.fn(async (url: string | URL | Request) => {
+      if (String(url).endsWith("/oauth.v2.user.access")) {
+        return Response.json({
           ok: true,
-          access_token: "bot-access",
-          refresh_token: "bot-refresh",
-          token_type: "bot",
+          access_token: "user-access",
+          refresh_token: "user-refresh",
+          token_type: "user",
           expires_in: 43_200,
-          scope: "channels:read,chat:write",
           authed_user: {
-            access_token: "user-access",
-            refresh_token: "user-refresh",
-            token_type: "user",
-            expires_in: 43_200,
-            scope: "search:read",
+            id: "U123",
+            scope: "channels:read,chat:write,search:read",
           },
-        }),
-      ),
-    );
+        });
+      }
+      return Response.json({
+        ok: true,
+        access_token: "bot-access",
+        refresh_token: "bot-refresh",
+        token_type: "bot",
+        expires_in: 43_200,
+        scope: "channels:read,chat:write",
+      });
+    });
+    vi.stubGlobal("fetch", fetcher);
 
-    const started = await services.flow.startAuthorization({ service: "slack" });
-    await services.flow.completeAuthorization({ state: started.state, code: "code" });
+    const userStarted = await services.flow.startAuthorization({ service: "slack" });
+    const userAuthorizationUrl = new URL(userStarted.authorizationUrl);
+    expect(userAuthorizationUrl.pathname).toBe("/oauth/v2_user/authorize");
+    expect(userAuthorizationUrl.searchParams.get("scope")).toContain("search:read");
+    await services.flow.completeAuthorization({ state: userStarted.state, code: "user-code" });
 
-    const credential = await services.connections.getCredential("slack");
-    expect(credential).toMatchObject({
+    const botStarted = await services.flow.startAuthorization({ service: "slackbot" });
+    const botAuthorizationUrl = new URL(botStarted.authorizationUrl);
+    expect(botAuthorizationUrl.pathname).toBe("/oauth/v2/authorize");
+    expect(botAuthorizationUrl.searchParams.get("scope")).not.toContain("search:read");
+    await services.flow.completeAuthorization({ state: botStarted.state, code: "bot-code" });
+
+    await expect(services.connections.getCredential("slack")).resolves.toMatchObject({
       authType: "oauth2",
-      accessToken: "bot-access",
+      accessToken: "user-access",
+      refreshToken: "user-refresh",
       tokenType: "Bearer",
-      providerSecret: {
-        userGrant: {
-          accessToken: "user-access",
-          refreshToken: "user-refresh",
-          scopes: ["search:read"],
-        },
+      profile: {
+        grantedScopes: ["channels:read", "chat:write", "search:read"],
       },
     });
-    expect(credential?.authType === "oauth2" ? credential.metadata : {}).not.toHaveProperty("authed_user");
+    await expect(services.connections.getCredential("slackbot")).resolves.toMatchObject({
+      authType: "oauth2",
+      accessToken: "bot-access",
+      refreshToken: "bot-refresh",
+      tokenType: "Bearer",
+      profile: {
+        grantedScopes: ["channels:read", "chat:write"],
+      },
+    });
+    expect(fetcher.mock.calls.map(([url]) => String(url))).toEqual([
+      "https://slack.com/api/oauth.v2.user.access",
+      "https://slack.com/api/oauth.v2.access",
+    ]);
+    const userCredential = await services.connections.getCredential("slack");
+    expect(userCredential?.authType === "oauth2" ? userCredential.metadata : {}).not.toHaveProperty("authed_user");
   });
 
   it("rejects expired OAuth authorization states", async () => {
