+++ src/providers/slack/executors.test.ts
@@ -1,0 +1,45 @@
+import type { ExecutionContext, ResolvedCredential } from "../../core/types.ts";
+
+import { describe, expect, it } from "vitest";
+import { executors as slackbotExecutors } from "../slackbot/executors.ts";
+import { executors as slackExecutors } from "./executors.ts";
+
+type OAuthCredential = Extract<ResolvedCredential, { authType: "oauth2" }>;
+
+describe("Slack authorization paths", () => {
+  it.each([
+    { actionId: "slack.list_channels", rawTokenType: "bot", execute: slackExecutors["slack.list_channels"]! },
+    {
+      actionId: "slackbot.list_channels",
+      rawTokenType: "user",
+      execute: slackbotExecutors["slackbot.list_channels"]!,
+    },
+  ])("rejects the other authorization path for $actionId", async ({ rawTokenType, execute }) => {
+    const context: ExecutionContext = {
+      getCredential: async () => oauthCredential(rawTokenType),
+    };
+
+    await expect(execute({}, context)).resolves.toMatchObject({
+      ok: false,
+      error: {
+        code: "authorization_failed",
+      },
+    });
+  });
+});
+
+function oauthCredential(rawTokenType: string): OAuthCredential {
+  return {
+    authType: "oauth2",
+    accessToken: "access-token",
+    tokenType: "Bearer",
+    profile: {
+      accountId: "U123",
+      displayName: "Example workspace",
+      grantedScopes: [],
+    },
+    metadata: {
+      rawTokenType,
+    },
+  };
+}
