+++ src/providers/slack/scopes.ts
@@ -14,6 +14,6 @@ export const slackReadScopes: string[] = [
 
 export const slackWriteScopes: string[] = ["chat:write", "im:write", "files:write", "reactions:write"];
 
-export const slackOAuthScopes: string[] = [...slackReadScopes, ...slackWriteScopes];
+export const slackBotOAuthScopes: string[] = [...slackReadScopes, ...slackWriteScopes];
 
-export const slackUserOAuthScopes: string[] = ["search:read"];
+export const slackUserOAuthScopes: string[] = [...slackReadScopes, ...slackWriteScopes, "search:read"];
